Variety changes contents && manage.py changepassword workaround implemented.
Now it shows the actual variety when selecting them thus it won't set the editing flag accidentally. So that's finally done.

Also since manage.py seems to want to _only_ set the salt to 12 characters(for reasons) instead of just letting the hasher handle it or using the salt that was previously set I've had to put a temporary stopgap in here. It'll be removed for the final version though.