Going to try to modify the code some more to have it update the challenge data.
So that it's fully refreshless on the admin page.
Then just comes modifying the padding characters for the hill cipher to use a subset of characters. Folllowed by finally doing the admin solves page.
Then a simpler use control panel(just showing their solves and also letting them update their password) and it'll be done finally.